Works a little better now (up until for statement)
Big issue is antlr grammar has no way of letting you put python methods or variables into lexer in a way that allows them to be called or seen by the rest of the lexer. You have to mangle it yourself for it to work. INDENTS and DEDENTS seem to break when dedenting more than once (parsing breaks after second-to-last for statement).